+++ lib/presentation/components/bottom_menu/bottom_menu.dart
@@ -1,7 +1,12 @@
+import 'dart:async';
+
+import 'package:birds/domain/datasources/config_source.dart';
 import 'package:birds/generated/l10n.dart';
 import 'package:birds/presentation/blocs/main_router.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:get_it/get_it.dart';
+import 'package:url_launcher/url_launcher_string.dart';
 
 @immutable
 class BottomMenu extends StatefulWidget {
@@ -21,14 +26,24 @@ class _BottomMenuState extends State<BottomMenu> {
   }
 
   void _handleChange(int index) {
-    if (index == 0) {
-      BlocProvider.of<MainRouterBLoC>(context).add(const MainRouterEvent.video());
-    } else {
-      BlocProvider.of<MainRouterBLoC>(context).add(const MainRouterEvent.about());
+    switch (index) {
+      case 2:
+        unawaited(launchUrlString(GetIt.I<ConfigSource>().donateUrl));
+        break;
+      case 1:
+        BlocProvider.of<MainRouterBLoC>(context).add(const MainRouterEvent.about());
+        setState(() {
+          _current = index;
+        });
+        break;
+      case 0:
+      default:
+        BlocProvider.of<MainRouterBLoC>(context).add(const MainRouterEvent.video());
+        setState(() {
+          _current = index;
+        });
+        break;
     }
-    setState(() {
-      _current = index;
-    });
   }
 
   @override
@@ -56,6 +71,16 @@ class _BottomMenuState extends State<BottomMenu> {
             color: MediaQuery.platformBrightnessOf(context) == Brightness.dark ? Colors.white : Colors.black,
           ),
           label: S.of(context).nav_about,
+        ),
+        BottomNavigationBarItem(
+          icon: Image.asset(
+            'assets/png/donate.png',
+            semanticLabel: 'Donate',
+            width: 32,
+            height: 32,
+            // color: MediaQuery.platformBrightnessOf(context) == Brightness.dark ? Colors.white : Colors.black,
+          ),
+          label: S.of(context).nav_donate,
         ),
       ],
     );