+++ lib/presentation/sceens/video/chat/chat_form_view.dart
@@ -13,7 +13,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_appauth/flutter_appauth.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
-import 'package:url_launcher/url_launcher_string.dart';
 
 @immutable
 class ChatFormView extends StatefulWidget {
@@ -50,27 +49,13 @@ class _ChatFormViewState extends State<ChatFormView> {
     }
   }
 
-  void _handleDonate() {
-    unawaited(launchUrlString('https://www.tbank.ru/cf/5mfwO0VNFF9'));
-  }
-
   @override
   Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        ElevatedButton(
-          onPressed: _handleDonate,
-          style: ElevatedButton.styleFrom(backgroundColor: const Color(0xffF08130)),
-          child: Text(S.of(context).donate),
-        ),
-        BlocBuilder<AuthBLoC, AuthState>(
-          builder: (context, state) => switch (state) {
-            OnAuthState(:UserEntity user) => ChatForm(user: user),
-            _ => ElevatedButton(onPressed: () => unawaited(_handleSign()), child: Text(S.of(context).sign_in)),
-          },
-        ),
-      ],
+    return BlocBuilder<AuthBLoC, AuthState>(
+      builder: (context, state) => switch (state) {
+        OnAuthState(:UserEntity user) => ChatForm(user: user),
+        _ => ElevatedButton(onPressed: () => unawaited(_handleSign()), child: Text(S.of(context).sign_in)),
+      },
     );
   }
 }